chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
requests==2.31.0
textblob==0.17.1
nltk==3.8.1
pydantic==2.5.0 